ExtractionTestAnswerGenerator extend AnswerGenerator
